New step file and board dimensions
I had to modify the board dimensions to fit the case.  I also had to move one of the RF connectors to get it out of the wall.  They're spaced okay now.  The panel design is based on this spacing, so if it's changed, we will have to modify the panel.