Modified schematics and placement
I removed some hardware that we were not really using to simplify the routing.  I also swapped pins on the usb protectors for the same reason.  I modified some components either to give them a real part number or to make them match (especially the footprints) of the satellite radio board.  I also fixed a couple of schematic errors (we had conflicting 5V busses).
I then moved the connectors to simplify routing as well, and moved a couple of blocks (nice work Miguel in putting the right parts into groups).
Finally, I completed the design of the supply input filter.  It's probably overkill, but can't hurt (even though the parts are huge!)